django_backend/requirements.txt
# Nauttec Yacht Platform - Django Backend Requirements
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-oauth-toolkit==1.7.1
django-allauth==0.57.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
django-filter==23.3
djoser==2.2.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
gunicorn==21.2.0
whitenoise==6.6.0